// src/api/openapi/models/UserAddressResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Country } from './Country';
import {
    CountryFromJSON,
    CountryFromJSONTyped,
    CountryToJSON,
    CountryToJSONTyped,
} from './Country';

/**
 * 
 * @export
 * @interface UserAddressResponseDto
 */
export interface UserAddressResponseDto {
    /**
     * 
     * @type {Country}
     * @memberof UserAddressResponseDto
     */
    country?: Country;
    /**
     * 
     * @type {string}
     * @memberof UserAddressResponseDto
     */
    city: string;
    /**
     * 
     * @type {string}
     * @memberof UserAddressResponseDto
     */
    psc: string;
    /**
     * 
     * @type {string}
     * @memberof UserAddressResponseDto
     */
    street: string;
    /**
     * 
     * @type {string}
     * @memberof UserAddressResponseDto
     */
    houseNumber: string;
}



/**
 * Check if a given object implements the UserAddressResponseDto interface.
 */
export function instanceOfUserAddressResponseDto(value: object): value is UserAddressResponseDto {
    if (!('city' in value) || value['city'] === undefined) return false;
    if (!('psc' in value) || value['psc'] === undefined) return false;
    if (!('street' in value) || value['street'] === undefined) return false;
    if (!('houseNumber' in value) || value['houseNumber'] === undefined) return false;
    return true;
}

export function UserAddressResponseDtoFromJSON(json: any): UserAddressResponseDto {
    return UserAddressResponseDtoFromJSONTyped(json, false);
}

export function UserAddressResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserAddressResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'country': json['country'] == null ? undefined : CountryFromJSON(json['country']),
        'city': json['city'],
        'psc': json['psc'],
        'street': json['street'],
        'houseNumber': json['houseNumber'],
    };
}

export function UserAddressResponseDtoToJSON(json: any): UserAddressResponseDto {
    return UserAddressResponseDtoToJSONTyped(json, false);
}

export function UserAddressResponseDtoToJSONTyped(value?: UserAddressResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'country': CountryToJSON(value['country']),
        'city': value['city'],
        'psc': value['psc'],
        'street': value['street'],
        'houseNumber': value['houseNumber'],
    };
}